Rename cleanupLeftGuilds test
