+++ src/config/sync.ts
@@ -1,2 +1,3 @@
-export const messageSyncLimit = process.env['MESSAGE_SYNC_LIMIT'] ? +process.env['MESSAGE_SYNC_LIMIT'] : 50_000
-
+export const messageSyncLimit = process.env['MESSAGE_SYNC_LIMIT']
+    ? +process.env['MESSAGE_SYNC_LIMIT']
+    : 50_000
